Add files via upload
This is the Amazon Delivery Analysis Project there were many factors which affect the delivery. Made this project to analyze the delivery performance and also analyze the agent performance. Where python used for cleaning and analyzing the delivery services, applied Machine Learning algorithm for better analysis. Create the comprehensive dashboard in Power BI.